feat: initial extraction of selection.rs
